Fix some warnings and incompatibilities with Visual C++

+++ lib/flatten.cpp
@@ -115,11 +115,11 @@ namespace MiniZinc {
   };
 
   void dumpEEb(const std::vector<EE>& ee) {
-    for (int i=0; i<ee.size(); i++)
+    for (unsigned int i=0; i<ee.size(); i++)
       std::cerr << *ee[i].b();
   }
   void dumpEEr(const std::vector<EE>& ee) {
-    for (int i=0; i<ee.size(); i++)
+    for (unsigned int i=0; i<ee.size(); i++)
       std::cerr << *ee[i].r();
   }
   std::vector<Expression*> toExpVec(std::vector<KeepAlive>& v) {
@@ -1451,7 +1451,7 @@ namespace MiniZinc {
       idx[i]=i;
     }
     std::sort(idx.begin(),idx.end(),CmpExpIdx(x));
-    int ci = 0;
+    unsigned int ci = 0;
     for (; ci<x.size(); ci++) {
       if (Lit* il = x[idx[ci]]()->dyn_cast<Lit>()) {
         d += c[idx[ci]]*il->v();
@@ -2140,15 +2140,16 @@ namespace MiniZinc {
                 if (isv->size() != 1)
                   throw FlatteningError(ti->loc(),"invalid array index set");
                 asize *= (isv->max(0)-isv->min(0)+1);
-                dims.push_back(std::pair<int,int>(isv->min(0).toInt(),isv->max(0).toInt()));
+                dims.push_back(std::pair<int,int>(static_cast<int>(isv->min(0).toInt()),
+					          static_cast<int>(isv->max(0).toInt())));
               }
             }
             Type tt = vd->ti()->type();
             tt._dim = 0;
             TypeInst* vti = new TypeInst(Location(),tt,vd->ti()->domain());
             
-            std::vector<Expression*> elems(asize.toInt());
-            for (int i=0; i<asize; i++) {
+            std::vector<Expression*> elems(static_cast<int>(asize.toInt()));
+            for (int i=0; i<static_cast<int>(asize.toInt()); i++) {
               VarDecl* nvd = new VarDecl(vd->loc(),vti,env.genId());
               nvd->introduced(vd->introduced());
               EE root_vd = flat_exp(env,Ctx(),nvd,NULL,constants().var_true);
@@ -2345,7 +2346,7 @@ namespace MiniZinc {
           std::vector<Expression*> in(c->n_generators());
           std::vector<Expression*> where;
           GCLock lock;
-          for (unsigned int i=0; i<c->n_generators(); i++) {
+          for (int i=0; i<c->n_generators(); i++) {
             if (c->in(i)->type().isvar()) {
               std::vector<Expression*> args(1);
               args[0] = c->in(i);
@@ -2353,7 +2354,7 @@ namespace MiniZinc {
               ub->type(Type::parsetint());
               ub->decl(env.orig->matchFn(ub));
               in[i] = ub;
-              for (unsigned int j=0; j<c->n_decls(i); j++) {
+              for (int j=0; j<c->n_decls(i); j++) {
                 BinOp* bo = new BinOp(Location(),c->decl(i,j)->id(), BOT_IN, c->in(i));
                 bo->type(Type::varbool());
                 where.push_back(bo);
@@ -2368,9 +2369,9 @@ namespace MiniZinc {
               gs._w = c->where();
             else
               where.push_back(c->where());
-            for (unsigned int i=0; i<c->n_generators(); i++) {
+            for (int i=0; i<c->n_generators(); i++) {
               std::vector<VarDecl*> vds(c->n_decls(i));
-              for (unsigned int j=0; j<c->n_decls(i); j++)
+              for (int j=0; j<c->n_decls(i); j++)
                 vds[i] = c->decl(i, j);
               gs._g.push_back(Generator(vds,in[i]));
             }
@@ -3862,7 +3863,7 @@ namespace MiniZinc {
               reallyFlat->addAnnotation(constants().ann.output_var);
             } else {
               std::vector<Expression*> args(reallyFlat->e()->type().dim());
-              for (int i=0; i<args.size(); i++) {
+              for (unsigned int i=0; i<args.size(); i++) {
                 if (nvi->e()->ti()->ranges()[i]->domain() == NULL) {
                   args[i] = new SetLit(Location(), eval_intset(reallyFlat->ti()->ranges()[i]->domain()));
                 } else {
@@ -3971,7 +3972,7 @@ namespace MiniZinc {
                       vd->flat()->addAnnotation(constants().ann.output_var);
                     } else {
                       std::vector<Expression*> args(vd->type().dim());
-                      for (int i=0; i<args.size(); i++) {
+                      for (unsigned int i=0; i<args.size(); i++) {
                         if (vd->ti()->ranges()[i]->domain() == NULL) {
                           args[i] = new SetLit(Location(), eval_intset(vd->flat()->ti()->ranges()[i]->domain()));
                         } else {
@@ -4154,7 +4155,7 @@ namespace MiniZinc {
                   }
                   if (needOutputAnn) {
                     std::vector<Expression*> args(vdi->e()->type().dim());
-                    for (int i=0; i<args.size(); i++) {
+                    for (unsigned int i=0; i<args.size(); i++) {
                       if (vdi->e()->ti()->ranges()[i]->domain() == NULL) {
                         args[i] = new SetLit(Location(), eval_intset(vdi->e()->flat()->ti()->ranges()[i]->domain()));
                       } else {
@@ -4304,7 +4305,7 @@ namespace MiniZinc {
     
     std::vector<VarDecl*> deletedVarDecls;
     while (startItem <= endItem) {
-      for (unsigned int i=startItem; i<=endItem; i++) {
+      for (int i=startItem; i<=endItem; i++) {
         VarDeclI* vdi = m[i]->dyn_cast<VarDeclI>();
         bool keptVariable = true;
         if (vdi!=NULL && !isOutput(vdi->e()) && env.vo.occurrences(vdi->e())==0 ) {
@@ -4372,7 +4373,7 @@ namespace MiniZinc {
           }
         }
       }
-      for (unsigned int i=startItem; i<=endItem; i++) {
+      for (int i=startItem; i<=endItem; i++) {
         if (VarDeclI* vdi = m[i]->dyn_cast<VarDeclI>()) {
           VarDecl* vd = vdi->e();
           if (!vdi->removed() && vd->e()) {
@@ -5006,7 +5007,7 @@ namespace MiniZinc {
     int msize = m->size();
     UNORDERED_NAMESPACE::unordered_set<Item*> globals;
     std::vector<int> declsWithIds;
-    for (unsigned int i=0; i<msize; i++) {
+    for (int i=0; i<msize; i++) {
       if ((*m)[i]->removed())
         continue;
       if (VarDeclI* vdi = (*m)[i]->dyn_cast<VarDeclI>()) {
@@ -5025,7 +5026,7 @@ namespace MiniZinc {
         
         if (vd->e() && vd->e()->isa<Id>()) {
           declsWithIds.push_back(i);
-          vdi->e()->payload(-i-1);
+          vdi->e()->payload(-static_cast<int>(i)-1);
         } else {
           vdi->e()->payload(i);
         }
